+++ src/functions/auto-pull/helpers/check-logs.js
