+++ src/functions/auto-pull/helpers/info-prompter.js
