+++ src/functions/clone/helpers/git-process.js
