+++ src/functions/clone/index.js
