+++ src/functions/config/helpers/set-global-config.js
