+++ src/functions/generate-gpg/helpers/delete-genkey.js
