+++ src/functions/generate-gpg/helpers/generate-gpg.js
