+++ src/functions/generate-gpg/helpers/info-prompter.js
