+++ src/functions/generate-gpg/helpers/init-shell-update.js
