+++ src/functions/generate-gpg/helpers/user-info.js
