+++ src/functions/generate-gpg/static/index.html
