+++ src/functions/generate-ssh/static/index.html
