+++ src/functions/push/helpers/git-process/stage.js
