+++ src/functions/push/helpers/sync-config.js
