+++ src/functions/version/index.js
@@ -1,6 +1,6 @@
 module.exports = (args) => {
   const fs = require('fs')
-  const pjson = require('../../package.json')
+  const pjson = require('../../../package.json')
   const clog = require('../../utils/loggers/console-log.js')
   clog.info(`Your currently installed git-assist version is ${pjson.version}`)
 }
