+++ src/utils/shell/get-init-file.js
